feat/#38: Fcm Repsotiry 설정

+++ data/src/main/java/project/graduation/crowd_sourcing/data/request/fcm/FcmWorkRequest.kt
@@ -1,6 +1,6 @@
 package project.graduation.crowd_sourcing.data.request.fcm
 
 data class FcmWorkRequest(
-    val workId: Long,
-    val memberId: Long
+    val workId: Int,
+    val memberId: Int
 )
